Missing namespace in data source is detected when setting
When applying settings from a namespace within a data source, the
previous behavior when said namespace wasn't found in the data was
confusing. Suppose the following data source and class:

    example_data_source = {
      some_namespace: {
        host: "localhost",
        port: 2113
      }
    }

    example_settings = Settings.build example_data_source

    class SomeClass
      setting :host
      setting :port
    end

If you attempted to supply an instance of `SomeClass` with
`example_settings`, and got the namespace wrong, the previous behavior
was a `NoMethodError` for `#each` on `NilClass`.

    # This would raise NoMethodError #each on NilClass
    example_settings.set(SomeClass.new, :wrong_namespace)

This commit causes the same code to raise a `Settings::Error` with a
message indicating that the namespace `:wrong_namespace` was not found
in the data source.